Prebuilt Swing Widgets to ease GUI rendering
The most important class here is TracerPanel, which renders
a button to start the rendering process, a canvas where the
image is rendered and a progress bar indicating the status
of the process.

The SwingListener is a JTrace Listener (it's in fact the
only part of the code that 'talks' to JTrace).

The DrawablePanel is used a simple Java 2D canvas.